// waste/src/enums/bookingStatus.enum.ts
export enum BookingStatus 
{
    PENDING = 'pending',
    ASSIGNED = 'assigned',
    ACCEPTED = 'accepted',
    EN_ROUTE = 'en_route',
    PICKED_UP = 'picked_up',
    COMPLETED = 'completed',
    CANCELLED = 'cancelled',
    FAILED = 'failed',
    CONFIRMED = 'confirmed',
    DECLINED = 'declined',
    REASSIGNING = 'reassigning',
}